+++ dataviva/api/rais/models.py
@@ -13,10 +13,10 @@ import json
 
 
 class BaseRais(object):
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
 
     wage = db.Column(db.Numeric(16, 2))
-    num_jobs = db.Column(db.Integer(11))
+    num_jobs = db.Column(db.Integer())
     wage_avg = db.Column(db.Numeric(16, 2))
 
     wage_growth = db.Column(db.Float())
@@ -29,16 +29,16 @@ class BaseRais(object):
 class Yi(BaseRais, db.Model, AutoSerialize):
 
     __tablename__ = 'rais_yi'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     cnae_id = db.Column(db.String(5), db.ForeignKey(Cnae.id), primary_key=True)
     pci = db.Column(db.Float())
 
-    cbo_diversity = db.Column(db.Integer(11))
+    cbo_diversity = db.Column(db.Integer())
     cbo_diversity_eff = db.Column(db.Float())
-    bra_diversity = db.Column(db.Integer(11))
+    bra_diversity = db.Column(db.Integer())
     bra_diversity_eff = db.Column(db.Float())
 
-    cnae_id_len = db.Column(db.Integer(1))
+    cnae_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Yi %d.%s>' % (self.year, self.cnae_id)
@@ -49,12 +49,12 @@ class Yb_rais(BaseRais, db.Model, AutoSerialize):
     __tablename__ = 'rais_yb'
     bra_id = db.Column(db.String(8), db.ForeignKey(Bra.id), primary_key=True)
 
-    cnae_diversity = db.Column(db.Integer(11))
+    cnae_diversity = db.Column(db.Integer())
     cnae_diversity_eff = db.Column(db.Float())
-    cbo_diversity = db.Column(db.Integer(11))
+    cbo_diversity = db.Column(db.Integer())
     cbo_diversity_eff = db.Column(db.Float())
 
-    bra_id_len = db.Column(db.Integer(1))
+    bra_id_len = db.Column(db.Integer())
 
     eci = db.Column(db.Float())
     distance = db.Column(db.Float())
@@ -70,12 +70,12 @@ class Yo(BaseRais, db.Model, AutoSerialize):
     __tablename__ = 'rais_yo'
     cbo_id = db.Column(db.String(6), db.ForeignKey(Cbo.id), primary_key=True)
 
-    cnae_diversity = db.Column(db.Integer(11))
+    cnae_diversity = db.Column(db.Integer())
     cnae_diversity_eff = db.Column(db.Float())
-    bra_diversity = db.Column(db.Integer(11))
+    bra_diversity = db.Column(db.Integer())
     bra_diversity_eff = db.Column(db.Float())
 
-    cbo_id_len = db.Column(db.Integer(1))
+    cbo_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Yo %d.%s>' % (self.year, self.cbo_id)
@@ -97,8 +97,8 @@ class Ybi(BaseRais, db.Model, AutoSerialize):
     distance = db.Column(db.Float())
     opp_gain = db.Column(db.Float())
 
-    bra_id_len = db.Column(db.Integer(1))
-    cnae_id_len = db.Column(db.Integer(1))
+    bra_id_len = db.Column(db.Integer())
+    cnae_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Ybi %d.%s.%s>' % (self.year, self.bra_id, self.cnae_id)
@@ -107,7 +107,7 @@ class Ybi(BaseRais, db.Model, AutoSerialize):
 class Ybi_reqs(db.Model, AutoSerialize):
 
     __tablename__ = 'rais_ybi_required'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     bra_id = db.Column(db.String(8), db.ForeignKey(Bra.id), primary_key=True)
     cnae_id = db.Column(db.String(5), db.ForeignKey(Cnae.id), primary_key=True)
 
@@ -120,12 +120,12 @@ class Ybi_reqs(db.Model, AutoSerialize):
 class Ybo(BaseRais, db.Model, AutoSerialize):
 
     __tablename__ = 'rais_ybo'
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     bra_id = db.Column(db.String(8), db.ForeignKey(Bra.id), primary_key=True)
     cbo_id = db.Column(db.String(6), db.ForeignKey(Cbo.id), primary_key=True)
 
-    bra_id_len = db.Column(db.Integer(1))
-    cbo_id_len = db.Column(db.Integer(1))
+    bra_id_len = db.Column(db.Integer())
+    cbo_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Ybo %d.%s.%s>' % (self.year, self.bra_id, self.cbo_id)
@@ -137,8 +137,8 @@ class Yio(BaseRais, db.Model, AutoSerialize):
     cnae_id = db.Column(db.String(5), db.ForeignKey(Cnae.id), primary_key=True)
     cbo_id = db.Column(db.String(6), db.ForeignKey(Cbo.id), primary_key=True)
 
-    cnae_id_len = db.Column(db.Integer(1))
-    cbo_id_len = db.Column(db.Integer(1))
+    cnae_id_len = db.Column(db.Integer())
+    cbo_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Yio %d.%s.%s>' % (self.year, self.cnae_id, self.cbo_id)
@@ -151,9 +151,9 @@ class Ybio(BaseRais, db.Model, AutoSerialize):
     cnae_id = db.Column(db.String(5), db.ForeignKey(Cnae.id), primary_key=True)
     cbo_id = db.Column(db.String(6), db.ForeignKey(Cbo.id), primary_key=True)
 
-    bra_id_len = db.Column(db.Integer(1))
-    cnae_id_len = db.Column(db.Integer(1))
-    cbo_id_len = db.Column(db.Integer(1))
+    bra_id_len = db.Column(db.Integer())
+    cnae_id_len = db.Column(db.Integer())
+    cbo_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Ybio %d.%s.%s.%s>' % (self.year, self.bra_id, self.cnae_id, self.cbo_id)
