+++ dataviva/api/sc/models.py
@@ -5,10 +5,10 @@ from dataviva.api.attrs.models import Bra, Course_sc, School
 class Sc(db.Model, AutoSerialize):
     __abstract__ = True
 
-    year = db.Column(db.Integer(4), primary_key=True)
+    year = db.Column(db.Integer(), primary_key=True)
     age = db.Column(db.Float())
-    classes = db.Column(db.Integer(11))
-    enrolled = db.Column(db.Integer(11))
+    classes = db.Column(db.Integer())
+    enrolled = db.Column(db.Integer())
     enrolled_growth = db.Column(db.Float())
     enrolled_growth_5 = db.Column(db.Float())
 
@@ -16,9 +16,9 @@ class Yb_sc(Sc):
     __tablename__ = 'sc_yb'
 
     bra_id = db.Column(db.String(9), db.ForeignKey(Bra.id), primary_key=True)
-    num_schools = db.Column(db.Integer(11))
+    num_schools = db.Column(db.Integer())
 
-    bra_id_len = db.Column(db.Integer(1))
+    bra_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Yb {0}.{1}>'.format(self.year, self.bra_id)
@@ -38,8 +38,7 @@ class Ybs(Sc):
 
     bra_id = db.Column(db.String(9), db.ForeignKey(Bra.id), primary_key=True)
     school_id = db.Column(db.String(8), db.ForeignKey(School.id), primary_key=True)
-
-    bra_id_len = db.Column(db.Integer(1))
+    bra_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Ybs %d.%s.%s>' % (self.year, self.bra_id, self.school_id)
@@ -50,8 +49,8 @@ class Ybc_sc(Sc):
     bra_id = db.Column(db.String(9), db.ForeignKey(Bra.id), primary_key=True)
     course_sc_id = db.Column(db.String(5), db.ForeignKey(Course_sc.id), primary_key=True)
 
-    bra_id_len = db.Column(db.Integer(1))
-    course_sc_id_len = db.Column(db.Integer(1))
+    bra_id_len = db.Column(db.Integer())
+    course_sc_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Ybc %d.%s.%s>' % (self.year, self.bra_id, self.course_sc_id)
@@ -61,7 +60,7 @@ class Yc_sc(Sc):
 
     course_sc_id = db.Column(db.String(5), db.ForeignKey(Course_sc.id), primary_key=True)
 
-    course_sc_id_len = db.Column(db.Integer(1))
+    course_sc_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Ybc %d.%s>' % (self.year, self.course_sc_id)
@@ -73,7 +72,7 @@ class Ysc(Sc):
     school_id = db.Column(db.String(8), db.ForeignKey(School.id), primary_key=True)
     course_sc_id = db.Column(db.String(5), db.ForeignKey(Course_sc.id), primary_key=True)
 
-    course_sc_id_len = db.Column(db.Integer(1))
+    course_sc_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Ysc %d.%s>' % (self.year, self.school_id)
@@ -86,8 +85,8 @@ class Ybsc(Sc):
     school_id = db.Column(db.String(8), db.ForeignKey(School.id), primary_key=True)
     course_sc_id = db.Column(db.String(5), db.ForeignKey(Course_sc.id), primary_key=True)
 
-    course_sc_id_len = db.Column(db.Integer(1))
-    bra_id_len = db.Column(db.Integer(1))
+    course_sc_id_len = db.Column(db.Integer())
+    bra_id_len = db.Column(db.Integer())
 
     def __repr__(self):
         return '<Ybsc %d.%s.%s.%s>' % (self.year, self.bra_id, self.school_id, self.course_sc_id)
